v7.5	fix(ffmpeg): Add hwupload filter for software-decoded .avi
Fixes a filter crash when transcoding .avi files.
When an .avi is software-decoded (CPU) and hardware-
encoded (VAAPI), we must now manually add a video
filter ('-vf format=nv12,hwupload') to upload the
CPU frame to the GPU for encoding. This resolves the
"Impossible to convert..." and "Error reinitializing filters!"
errors.